Move Application to application.py
