+++ python/ray/serve/tests/test_cli.py
@@ -12,7 +12,7 @@ import ray
 from ray import serve
 from ray.tests.conftest import tmp_working_dir  # noqa: F401, E501
 from ray._private.test_utils import wait_for_condition
-from ray.serve.api import Application
+from ray.serve.application import Application
 from ray.serve.deployment_graph import RayServeDAGHandle
